APPLE: fix daemon version checker (#2803)
* APPLE: fix daemon version checker

* daemon menubar quit improvements

* On disconect with reconnect skip disconnecting state macOS

* On quit imrovements macOS

* RC2 core

* New Crowdin translations by GitHub Action (#2778)


* Add iOS support for connecting when disconnect with reconnect

---------
