--- nym-vpn-apple/Services/Sources/Services/Constants/Constants.swift
@@ -1,21 +1,0 @@
-import Foundation
-
-public enum Constants: String {
-    case groupID = "group.net.nymtech.vpn"
-
-    case supportURL = "https://support.nym.com/hc"
-    case termsOfUseURL = "https://nym.com/vpn-terms"
-    case privacyPolicyURL = "https://nym.com/vpn-privacy-statement"
-    case newSupportRequest = "https://support.nym.com/hc/requests/new"
-    case operatorDocs = "https://nym.com/docs"
-    case downloadLink = "https://nym.com/download"
-    case createAccounURL = "https://nym.com/account/create"
-    case pricingURL = "https://nym.com/pricing"
-
-    case telegramLink = "https://nym.com/go/telegram"
-    case discordLink = "https://nym.com/go/discord"
-    case ghIssuesLink = "https://nym.com/go/github/nym-vpn-client/issues"
-    case matrixLink = "https://nym.com/go/matrix"
-
-    case logFileName = "Logs.log"
-}
