--- nym-vpn-apple/Services/Sources/Services/DarwinNotificationCenter/DarwinNotificationKey.swift
@@ -1,10 +1,0 @@
-import Foundation
-import Constants
-
-public enum DarwinNotificationKey: String {
-    case reconfigureLogs
-
-    public var key: String {
-        Constants.groupID.rawValue + self.rawValue
-    }
-}
