--- nym-vpn-apple/Services/Sources/Services/NymLogger/FileLogHandler.swift
@@ -1,73 +1,0 @@
-import Foundation
-import Logging
-import Constants
-
-public class FileLogHandler: LogHandler {
-    private let label: String
-    private let logFileManager: LogFileManager
-
-    public init(label: String, logFileManager: LogFileManager) {
-        self.label = label
-        self.logFileManager = logFileManager
-    }
-
-    public var metadata = Logging.Logger.Metadata()
-    public var logLevel = Logging.Logger.Level.info
-
-    public subscript(metadataKey key: String) -> Logging.Logger.Metadata.Value? {
-        get { metadata[key] }
-        set { metadata[key] = newValue }
-    }
-
-    // swiftlint:disable:next function_parameter_count
-    public func log(
-        level: Logger.Level,
-        message: Logger.Message,
-        metadata: Logger.Metadata?,
-        source: String,
-        file: String,
-        function: String,
-        line: UInt
-    ) {
-        var fullMetadata = self.metadata
-        if let metadata = metadata {
-            fullMetadata.merge(metadata) { $1 }
-        }
-
-        var metadataOutput = fullMetadata.formatted()
-        if !metadataOutput.isEmpty {
-            metadataOutput = " " + metadataOutput
-        }
-
-        let logLine = "\(Date()) [\(label)] \(level.emoji) \(level)\(metadataOutput): \(message)\n"
-        logFileManager.write(logLine)
-    }
-}
-
-extension Logging.Logger.Metadata {
-    func formatted() -> String {
-        map { key, value in "\(key)=\(value)" }
-            .joined(separator: " ")
-    }
-}
-
-extension Logging.Logger.Level {
-    var emoji: String {
-        switch self {
-        case .trace:
-            return "👀"
-        case .debug:
-            return "⌨️"
-        case .info:
-            return "ℹ"
-        case .notice:
-            return "📣"
-        case .warning:
-            return "⚠️"
-        case .error:
-            return "⛔️"
-        case .critical:
-            return "🔥"
-        }
-    }
-}
