--- nym-vpn-apple/Services/Sources/Services/NymLogger/LogFileType.swift
@@ -1,10 +1,0 @@
-public enum LogFileType: String, CaseIterable {
-#if os(iOS)
-    case app
-    case tunnel
-    case library
-#elseif os(macOS)
-    case app
-    case daemon
-#endif
-}
