+++ nym-vpn-apple/ServicesMutual/Sources/DarwinNotificationCenter/DarwinNotificationKey.swift
@@ -1,0 +1,10 @@
+import Foundation
+import Constants
+
+public enum DarwinNotificationKey: String {
+    case reconfigureLogs
+
+    public var key: String {
+        Constants.groupID.rawValue + self.rawValue
+    }
+}
