+++ nym-vpn-apple/ServicesMutual/Sources/NymLogger/LogFileType.swift
@@ -1,0 +1,10 @@
+public enum LogFileType: String, CaseIterable {
+#if os(iOS)
+    case app
+    case tunnel
+    case library
+#elseif os(macOS)
+    case app
+    case daemon
+#endif
+}
