+++ nym-vpn-apple/ServicesMutual/Sources/NymLogger/NymLogger.swift
@@ -1,0 +1,13 @@
+import Logging
+
+public class NymLogger {
+    public let logFileManager: LogFileManager
+
+    public init() {
+        let newLogFileManager = LogFileManager(logFileType: .app)
+        self.logFileManager = newLogFileManager
+        LoggingSystem.bootstrap { label in
+            FileLogHandler(label: label, logFileManager: newLogFileManager)
+        }
+    }
+}
