+++ nym-vpn-apple/ServicesMutual/Sources/TunnelStatus/AfterDisconnectAction.swift
@@ -1,0 +1,4 @@
+public enum AfterDisconnectAction: Int, Codable {
+    case reconnect
+    case offline
+}
